+++ docs/CHANGELOG.md
@@ -16,12 +16,14 @@
 - removed legacy pdoc-editor
 - frontend: improved/fixed desc-rendering
 - improve imageInlineLoader
+- added support for automatic nearby track pois
 
 ### new features
 - added optional password-encryption of static-viewer-version
 - frontend: open modal view on maps + lists
 - backend/frontend: add page-management
 - improve imageInlineLoader
+- added support for automatic nearby track pois
 
 ### improvements
 - frontend: improved filter-suggestion for poi-search